Centralize loadReferers
